+++ webextension/data/js/classes/ZDK.js
@@ -132,6 +132,12 @@ class ZDK{
 			}
 		}
 	}
+
+	/**
+	 * Turns a Map into a JSON object
+	 * @param {Map}  myMap
+	 * @returns {Object}
+	 */
 	static mapToObj(myMap){
 		if(myMap instanceof Map){
 			let obj = {};
@@ -145,6 +151,29 @@ class ZDK{
 	}
 
 
+	/**
+	 * Ellipse str if the str string length is higher than strMaxLength
+	 * @param {String} str
+	 * @param {Number} strMaxLength Max string length wanted
+	 * @returns {String} String of strMaxLength length or less
+	 */
+	static stringEllipse(str, strMaxLength) {
+		if(typeof str!=="string" || typeof strMaxLength!=="number"){
+			throw "Argument type problem";
+		}
+		if(str.length>strMaxLength){
+			return `${str.substring(0, strMaxLength-3)}...`
+		} else {
+			return str;
+		}
+	}
+
+
+	/**
+	 *
+	 * @param {Object} obj
+	 * @returns {Promise<HTMLImageElement>}
+	 */
 	loadImage(obj={}){
 		return new Promise((resolve, reject)=>{
 			let imgNode;
@@ -183,6 +212,12 @@ class ZDK{
 			}
 		})
 	}
+
+	/**
+	 *
+	 * @param {Blob} blob
+	 * @return {Promise}
+	 */
 	loadBlob(blob){
 		return new Promise((resolve, reject)=>{
 			const reader = new FileReader();
@@ -242,7 +277,7 @@ class ZDK{
 	 * @param action
 	 * @param selector
 	 * @param html
-	 * @param doc
+	 * @param {HTMLDocument} doc
 	 * @returns {Element}
 	 */
 	insertHtml(action, selector, html, doc=document){
@@ -275,7 +310,7 @@ class ZDK{
 	/**
 	 * @param selector
 	 * @param html
-	 * @param doc
+	 * @param {HTMLDocument} doc
 	 * @returns {Element}
 	 */
 	appendTo(selector, html, doc=document){
@@ -286,7 +321,7 @@ class ZDK{
 	 *
 	 * @param selector
 	 * @param html
-	 * @param doc
+	 * @param {HTMLDocument} doc
 	 * @returns {Element}
 	 */
 	insertBefore(selector, html, doc=document){
@@ -367,10 +402,22 @@ class Params extends Map {
 		return array.join('&');
 	}
 }
+
+/**
+ *
+ * @param {Object} options
+ * @returns {{get: get, post: post}}
+ * @constructor
+ */
 function Request(options){
 	if(typeof options.url !== "string" /*&& typeof options.onComplete !== "function"*/){
 		consoleMsg("warn", "Error in options");
 	} else {
+		/**
+		 *
+		 * @param {String} method
+		 * @returns {Promise<Object>}
+		 */
 		let core = function(method){
 			return new Promise(resolve=>{
 				let xhr;